agent: store agent-connetion at end of create instead of start
To aid in firecracker enabling, start of the guest is pushed later,
during the OCI "start" command handling.  It was assumed that in-memory
data for the agent state would be available at the time of call to
'startSandbox' for agent.

Necessary state information include parameters to connect to the guests'
agent.  In case of VSOCK, we have all the information after calling
agent's 'configure' function, and can store this to persistent storage
at the end of 'createSandbox'

Note, this isn't the case for non-vsock, who needs the PID of the proxy
to be included as well.  This won't be available until later in the
flow, so indeed, this patch is pretty dirty, hacky, and terrible.

BUT! But! it does enable appropriate agent state to be restored so a new
client connection can be made to the original CID/vsock url.

Signed-off-by: Eric Ernst <eric.ernst@intel.com>